fix(dcs-bios): widen fuelLbs return to uint32_t (max 99900 > uint16_t)
The fuel totalizer assembles three dial digits (10K, 1K, 100) into a
single value. Maximum possible: 9×10000 + 9×1000 + 9×100 = 99900.
uint16_t max is 65535 — silent truncation for fuel > 65900 lbs.
Changed return type to uint32_t in both header and implementation.

+++ include/dcs_bios.h
@@ -53,7 +53,7 @@ namespace DcsBios {
   bool    rwrMslLaunch();
   bool    storesConfigLight();
   uint8_t storesConfigSw();
-  uint16_t fuelLbs();
+  uint32_t fuelLbs();
   uint8_t  chaffCount();
   uint8_t  flareCount();
   bool     ecmTransmitting();
